+++ CHANGELOG.md
@@ -5,6 +5,7 @@ CHANGELOG
   - [PR #312](https://github.com/caxlsx/caxlsx/pull/312) Added 2D/flat PieChart drawing
   - [PR #317](https://github.com/caxlsx/caxlsx/pull/317) Apply style for columns without defining cells
   - [PR #345](https://github.com/caxlsx/caxlsx/pull/345) Show outline symbols by default to match original behavior
+  - [PR #334](https://github.com/caxlsx/caxlsx/pull/334) Add pattern fill options to add_style
 
 - **October.30.23**: 4.0.0
   - [PR #189](https://github.com/caxlsx/caxlsx/pull/189) **breaking** Make `Axlsx::escape_formulas` true by default to mitigate [Formula Injection](https://www.owasp.org/index.php/CSV_Injection) vulnerabilities.